Use Actor_MenuMan* type in several functions in Menu/radio.c

+++ src/Menu/menu_radio_8004D35C.c
@@ -5,5 +5,5 @@ extern Actor_MenuMan gMenuMan_800BD360;
 
 void menu_radio_8004D35C(void)
 {
-    sub_8004124C((int)&gMenuMan_800BD360);
+    sub_8004124C(&gMenuMan_800BD360);
 }
